backend/requirements.txt
# FastAPI
fastapi==0.110.1
uvicorn[standard]==0.29.0

# Authentification & sécurité
python-jose[cryptography]==3.3.0      # Pour signer les JWT
passlib[bcrypt]==1.7.4                 # Pour hasher les mots de passe

# ORM & base de données
sqlalchemy==2.0.30
psycopg2-binary==2.9.9                 # PostgreSQL driver (ou sqlite3 intégré)
alembic==1.13.1                        # Migration DB

# Données & validation
pydantic==2.7.3
pydantic[email]==2.7.3                # Pour la validation des emails
email-validator
python-multipart==0.0.9               # Pour parser les formulaires (upload, login)

# Gestion .env
python-dotenv==1.0.1

# Pour exporter/traiter du XLSX (import/export horaires)
openpyxl==3.1.2
pandas==2.2.2

# Tests
pytest==8.2.1
httpx==0.27.0

# (optionnel) Validation XLSX / formats
xlrd==2.0.1
xlsxwriter==3.2.0
